Add path and runtime-path variables.
Replace "dir:" and "file:" prefixes with a unified "path:" prefix.
The placeholder is substituted with absolute paths to the files in the
installation directory.
A new "runtime-path:" prefix is substituted with the file path as it
appears during command run-time. This may differ from the real path if
containers are being used.